refactor(snooze): a finding's key and anchor move to their own module
A pure move: no message, exit code, signature or file format changes,
and no test asserts anything new. `anchor_file`, `finding_keys` and
`snoozed_anchors` read the findings contract, not the index file and not
the CLI, so neither module they sat beside was their home.

Preparation for #163: the lapse rule is expressed in exactly these terms
and is about to be built here, and the move keeps snooze.py clear of the
200-line gate for what follows.
